# dgc-extracted/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DGC Arcade API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bets:
    post:
      operationId: placeBet
      tags: [bets]
      summary: Place a bet and resolve the round immediately
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BetInput"
      responses:
        "200":
          description: Bet result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BetResult"
        "400":
          description: Insufficient balance or validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    get:
      operationId: listBets
      tags: [bets]
      summary: Get the recent bet history for current user
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Recent bets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BetRecord"
  /bets/recent-all:
    get:
      operationId: listRecentBetsAll
      tags: [bets]
      summary: Public feed of recent bets across all users (wow endpoint)
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: Recent public bets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BetRecord"

  # ── Transactions ──────────────────────────────────────────────────────
  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List the authenticated user's transaction history
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
  /transactions/deposit/initiate:
    post:
      operationId: initiateDeposit
      tags: [transactions]
      summary: Create an OxaPay payment link for a deposit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepositInput"
      responses:
        "200":
          description: OxaPay checkout URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepositLink"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /transactions/deposit/callback:
    post:
      operationId: oxapayCallback
      tags: [transactions]
      summary: OxaPay IPN webhook — credit balance on confirmed payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OxapayCallbackInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /transactions/withdraw:
    post:
      operationId: requestWithdrawal
      tags: [transactions]
      summary: Request a withdrawal (queued for manual review)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawInput"
      responses:
        "200":
          description: Withdrawal queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Insufficient balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Leaderboard & Stats ───────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]

    # ── Auth ──────────────────────────────────────────────────────────
...
    BetInput:
      type: object
      required: [gameId, amount]
      properties:
        gameId:
          type: integer
        amount:
          type: number
        clientSeed:
          type: ["string", "null"]
        meta:
          type: ["object", "null"]
          description: Game-specific options (e.g. chosen side for coin flip)
          additionalProperties: {}
    BetResult:
      type: object
      required: [bet, newBalance, won, payout, multiplier]
      properties:
        bet:
          $ref: "#/components/schemas/BetRecord"
        newBalance:
          type: number
        won:
          type: boolean
        payout:
          type: number
        multiplier:
          type: number
    BetRecord:
      type: object
      required: [id, userId, username, gameId, gameName, amount, payout, won, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        username:
          type: string
        gameId:
          type: integer
        gameName:
          type: string
        amount:
          type: number
        payout:
          type: number
        won:
          type: boolean
        multiplier:
          type: ["number", "null"]
        serverSeed:
          type: ["string", "null"]
        clientSeed:
          type: ["string", "null"]
        meta:
          type: ["object", "null"]
          additionalProperties: {}
        createdAt:
          type: string

    # ── Transactions ──────────────────────────────────────────────────
    DepositInput:
      type: object
      required: [amount, currency]
      properties:
        amount:
          type: number
        currency:
          type: string
          description: "e.g. BTC, ETH, LTC, USDT"
    DepositLink:
      type: object
      required: [paymentUrl, trackId]
      properties:
        paymentUrl:
          type: string
        trackId:
          type: string
    OxapayCallbackInput:
      type: object
      properties:
        trackId:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        amount:
          type: ["number", "null"]
        currency:
          type: ["string", "null"]
        orderId:
          type: ["string", "null"]
    WithdrawInput:
      type: object
      required: [amount, currency, address]
      properties:
        amount:
          type: number
        currency:
          type: string
        address:
          type: string
    Transaction:
      type: object
      required: [id, userId, type, amount, currency, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
          enum: [deposit, withdrawal, bet_win, bet_loss]
        amount:
          type: number
        currency:
          type: string
        status:
          type: string
          enum: [pending, completed, failed]
        txHash:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        createdAt:
          type: string

    # ── Leaderboard ───────────────────────────────────────────────────
...